+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - improved messaging on plugin child bridge modal
 - further improvements to plugin search
 - implement and allow switching to scoped plugins
+- merge plugin info and verified modals
 
 ### Other Changes